Update TInterval to use float types instead of int
- Changed constructor and setter parameters from int to float
- Updated getter return types from int to float
- Improved PHP style by removing spaces before colons in return types
- Updated getCount() return type to float for consistency

+++ src/Types/TInterval.php
@@ -7,52 +7,52 @@ class TInterval
 	public $min;
 	public $max;
 
-	public function __construct(int $min, int $max)
+	public function __construct(float $min, float $max)
 	{
 		$this->setMin($min);
 		$this->setMax($max);
 	}
 
-	public function setMin(int $number) : TInterval
+	public function setMin(float $number): TInterval
 	{
 		$this->min = $number;
 
 		return $this;
 	}
 
-	public function getMin() : int
+	public function getMin(): float
 	{
 		return $this->min;
 	}
 
-	public function setMax(int $number) : TInterval
+	public function setMax(float $number): TInterval
 	{
 		$this->max = $number;
 
 		return $this;
 	}
 
-	public function getMax() : int
+	public function getMax(): float
 	{
 		return $this->max;
 	}
 
-	public function getRange() : array
+	public function getRange(): array
 	{
 		return range($this->getMin(), $this->getMax());
 	}
 
-	public function getCount() : int
+	public function getCount(): float
 	{
 		return count($this->getRange());
 	}
 
-	public function getIsSingle() : bool
+	public function getIsSingle(): bool
 	{
 		return $this->getCount() == 1;
 	}
 
-	public function getIsDouble() : bool
+	public function getIsDouble(): bool
 	{
 		return $this->getCount() == 2;
 	}
